requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-community
chromadb
sentence-transformers
langchain_ollama
ollama
pytest
httpx